refactor:структура проекта становится мультимодульная, добавлена логирование в exceptionHandler, обновил компонентные тесты, данные для тестов вынесены в отдельные json файлы. Установленно правило для jackson которое запрещает конвертировать разные типы в строку.
