+++ app/src/java/com/kavencore/moneyharbor/app/entity/Account.java
@@ -1,0 +1,53 @@
+package com.kavencore.moneyharbor.app.entity;
+
+import jakarta.persistence.*;
+import lombok.*;
+import org.hibernate.proxy.HibernateProxy;
+
+import java.math.BigDecimal;
+import java.util.Objects;
+import java.util.UUID;
+
+@Entity
+@Table(name = "accounts")
+@Getter
+@Setter
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+@ToString(onlyExplicitlyIncluded = true)
+public class Account {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.UUID)
+    @ToString.Include
+    private UUID id;
+
+    @ToString.Include
+    private String title;
+
+    @Enumerated(EnumType.STRING)
+    @ToString.Include
+    private Currency currency;
+
+    @ToString.Include
+    private BigDecimal amount;
+
+    @Override
+    public final boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null) return false;
+        Class<?> oEffectiveClass = o instanceof HibernateProxy ? ((HibernateProxy) o).getHibernateLazyInitializer().getPersistentClass() : o.getClass();
+        Class<?> thisEffectiveClass = this instanceof HibernateProxy ? ((HibernateProxy) this).getHibernateLazyInitializer().getPersistentClass() : this.getClass();
+        if (thisEffectiveClass != oEffectiveClass) return false;
+        Account account = (Account) o;
+        return getId() != null && Objects.equals(getId(), account.getId());
+    }
+
+    @Override
+    public final int hashCode() {
+        return this instanceof HibernateProxy
+                ? ((HibernateProxy) this).getHibernateLazyInitializer().getPersistentClass().hashCode()
+                : getClass().hashCode();
+    }
+}
