+++ app/src/resources/db/changelog/db.changelog-master.yaml
@@ -1,0 +1,3 @@
+databaseChangeLog:
+  - include:
+      file: db/changelog/001-init-accounts.sql
