+++ app/src/resources/static/swagger/openapi.yaml
