+++ component-test/src/test/resources/json/accounts/create-amount-integer-too-long.json
@@ -1,0 +1,5 @@
+{
+  "currency": "RUB",
+  "amount": 1000000000000000000.00,
+  "title": "Too big"
+}
