+++ component-test/src/test/resources/json/accounts/create-amount-scale-3.json
@@ -1,0 +1,5 @@
+{
+  "currency": "RUB",
+  "amount": 10.123,
+  "title": "Scale 3"
+}
