+++ component-test/src/test/resources/json/accounts/create-invalid-enum.json
@@ -1,0 +1,4 @@
+{
+  "currency": "EURO",
+  "amount": 10.00
+}
