+++ component-test/src/test/resources/json/accounts/create-missing-currency.json
@@ -1,0 +1,4 @@
+{
+  "amount": 10.00,
+  "title": "Без валюты"
+}
