+++ component-test/src/test/resources/json/accounts/create-ok.json
@@ -1,0 +1,5 @@
+{
+  "currency": "RUB",
+  "amount": 1500.50,
+  "title": "Зарплатный"
+}
