+++ component-test/src/test/resources/json/accounts/create-title-wrong-type.json
@@ -1,0 +1,5 @@
+{
+  "currency": "USD",
+  "amount": 10.00,
+  "title": 123
+}
