+++ component-test/src/test/resources/json/accounts/create-with-currency-not-in-enum.json
@@ -1,0 +1,4 @@
+{
+  "currency": "CZH",
+  "amount": 10.00
+}
