+++ component-test/src/test/resources/json/accounts/create-without-title.json
@@ -1,0 +1,4 @@
+{
+  "currency": "USD",
+  "amount": 0.00
+}
