+++ settings.gradle.kts
@@ -1,1 +1,4 @@
 rootProject.name = "MoneyHarbor"
+
+include("app")
+include("component-test")
