--- src/main/java/com/kavencore/moneyharbor/app/entity/Currency.java
@@ -1,6 +1,0 @@
-package com.kavencore.moneyharbor.app.entity;
-
-/**
- * Код валюты по ISO 4217.
- */
-public enum Currency { BYN, RUB, USD }
