--- src/main/java/com/kavencore/moneyharbor/app/infrastructure/exceptionhandler/GlobalExceptionHandler.java
@@ -1,101 +1,0 @@
-package com.kavencore.moneyharbor.app.infrastructure.exceptionhandler;
-
-import com.kavencore.moneyharbor.app.infrastructure.exception.AccountNotFoundException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.validation.ConstraintViolationException;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ProblemDetail;
-import org.springframework.http.ResponseEntity;
-import org.springframework.http.converter.HttpMessageNotReadableException;
-import org.springframework.web.ErrorResponseException;
-import org.springframework.web.bind.MethodArgumentNotValidException;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-import org.springframework.web.method.annotation.MethodArgumentTypeMismatchException;
-
-import java.net.URI;
-
-@RestControllerAdvice
-public class GlobalExceptionHandler {
-
-    /**
-     * 404 Not Found — только для GET /accounts/{id}
-     */
-    @ExceptionHandler(AccountNotFoundException.class)
-    public ResponseEntity<ProblemDetail> handleNotFound(AccountNotFoundException ex, HttpServletRequest req) {
-        var pd = ProblemDetail.forStatus(HttpStatus.NOT_FOUND);
-        pd.setTitle("Not Found");
-        pd.setDetail(ex.getMessage());
-        pd.setInstance(URI.create(req.getRequestURI()));
-        return ResponseEntity.status(HttpStatus.NOT_FOUND).body(pd);
-    }
-
-    /**
-     * 400 Bad Request — невалидный UUID в path {id}
-     */
-    @ExceptionHandler(MethodArgumentTypeMismatchException.class)
-    public ResponseEntity<ProblemDetail> handleTypeMismatch(MethodArgumentTypeMismatchException ex, HttpServletRequest req) {
-        var pd = ProblemDetail.forStatus(HttpStatus.BAD_REQUEST);
-        pd.setTitle("Bad Request");
-        pd.setDetail("Invalid UUID in path parameter 'id'");
-        pd.setInstance(URI.create(req.getRequestURI()));
-        return ResponseEntity.badRequest().body(pd);
-    }
-
-    /**
-     * 400 Bad Request — битый JSON / ошибка десериализации
-     */
-    @ExceptionHandler(HttpMessageNotReadableException.class)
-    public ResponseEntity<ProblemDetail> handleBadJson(HttpMessageNotReadableException ex, HttpServletRequest req) {
-        var pd = ProblemDetail.forStatus(HttpStatus.BAD_REQUEST);
-        pd.setTitle("Bad Request");
-        pd.setDetail("Invalid value in request body");
-        pd.setInstance(URI.create(req.getRequestURI()));
-        return ResponseEntity.badRequest().body(pd);
-    }
-
-    /**
-     * 400 Bad Request — Bean Validation на DTO (@Valid)
-     */
-    @ExceptionHandler(MethodArgumentNotValidException.class)
-    public ResponseEntity<ProblemDetail> handleValidation(MethodArgumentNotValidException ex, HttpServletRequest req) {
-        var pd = ProblemDetail.forStatus(HttpStatus.BAD_REQUEST);
-        pd.setTitle("Bad Request");
-        pd.setDetail("Validation failed");
-        pd.setInstance(URI.create(req.getRequestURI()));
-        var errors = ex.getBindingResult().getFieldErrors().stream()
-                .map(fe -> String.format("%s: %s", fe.getField(), fe.getDefaultMessage()))
-                .toList();
-        pd.setProperty("errors", errors);
-        return ResponseEntity.badRequest().body(pd);
-    }
-
-    /**
-     * 400 Bad Request — нарушения Bean Validation при валидации параметров
-     * (например, @PathVariable, @RequestParam или @Validated-методы сервисов).
-     */
-    @ExceptionHandler(ConstraintViolationException.class)
-    public ResponseEntity<ProblemDetail> handleConstraint(ConstraintViolationException ex, HttpServletRequest req) {
-        var pd = ProblemDetail.forStatus(HttpStatus.BAD_REQUEST);
-        pd.setTitle("Bad Request");
-        pd.setDetail("Validation failed");
-        pd.setInstance(URI.create(req.getRequestURI()));
-        var errors = ex.getConstraintViolations().stream()
-                .map(v -> v.getPropertyPath() + ": " + v.getMessage())
-                .toList();
-        pd.setProperty("errors", errors);
-        return ResponseEntity.badRequest().body(pd);
-    }
-
-    /**
-     * На всякий случай: пробрасываем ProblemDetail из ErrorResponseException как есть
-     */
-    @ExceptionHandler(ErrorResponseException.class)
-    public ResponseEntity<ProblemDetail> handleErrorResponse(ErrorResponseException ex, HttpServletRequest req) {
-        var body = ex.getBody();
-        if (body.getInstance() == null) {
-            body.setInstance(URI.create(req.getRequestURI()));
-        }
-        return ResponseEntity.status(ex.getStatusCode()).body(body);
-    }
-}
