--- src/main/resources/db/changelog/001-init-accounts.yaml
@@ -1,33 +1,0 @@
-databaseChangeLog:
-  - changeSet:
-      id: 001-init-accounts
-      author: you
-      changes:
-        - createTable:
-            tableName: accounts
-            columns:
-              - column:
-                  name: id
-                  type: UUID
-                  constraints:
-                    primaryKey: true
-                    nullable: false
-              - column:
-                  name: title
-                  type: TEXT
-                  constraints:
-                    nullable: false
-              - column:
-                  name: currency
-                  type: TEXT
-                  constraints:
-                    nullable: false
-              - column:
-                  name: amount
-                  type: NUMERIC(19,2)
-                  constraints:
-                    nullable: false
-
-      rollback:
-        - dropTable:
-            tableName: accounts
